mira_client_pi/requirements.txt
PyQt5
requests
sounddevice
gtts
google-cloud-speech